change files and add connection gui
