--- gui_events.py
@@ -1,2 +1,0 @@
-def send_message(event):
-    print(input_str.get())
