--- refactoring/client.py
@@ -1,4 +1,0 @@
-from Root import Root
-
-if __name__ == '__main__':
-    Root().mainloop()
